[ui] Allow overriding run tags when performing re-execution (#28875)
## Summary & Motivation


https://linear.app/dagster-labs/issue/OPER-1624/override-run-tags-when-performing-a-re-execution-from-failure

This PR exposes the `extra_tags` option on re-execution (already
available in the python API) via graphql, and allows the user to put in
additional tags when re-executing using one of the two core strategies
(ALL_STEPS and FROM_FAILURE)

We allow re-execution from a bunch of places, and I put the new UI into
the `ReexecutionDialog` which is used when you select multiple runs on
the runs feed view and chose the bulk action. For other re-execute menu
items, including the individual run actions, the job page "reexecute
last run" option, and the run details page options, I added menu item
tooltips and made shift-clicking open the dialog so you can add tags
there too.

**Q: Why not show all the existing tags and let the user edit them?**

This was what I tried at first. Adding tags to a re-execution overrides
existing tags and allows you to pass new tags, but it cannot be used to
remove a tag from the parent run. That can be done in the Python API by
setting `use_parent_run_tags=False` and re-specifying all tags that are
not `TAGS_TO_MAYBE_OMIT_ON_RETRY`, but the behavior is complicated. I
chose not to expose this option to re-execute /without/ some tags
because 1) it wasn't asked for and 2) it would require the UI to either
ask the user to put in all their tags from scratch, or be aware of
exactly what tags from the parent run would have normally used and
should be specified in `extra_tags` along with
`use_parent_run_tags=False`.

The backend API is specifically for for adding "extra tags", and after
investigating I don't think we should try to build a "review + modify"
style UI in the front-end.

## How I Tested These Changes

I tested this manually and by adding new tests! 


![image](https://github.com/user-attachments/assets/1ae6f7b0-52e6-416d-a337-b0deb3c16af7)


![image](https://github.com/user-attachments/assets/a2c9bf20-df9e-41b9-aa0f-cdbdfbd5ef0f)


![image](https://github.com/user-attachments/assets/668e7d17-4f00-45bd-b673-76ca5c3306ca)


![image](https://github.com/user-attachments/assets/cbdf3c20-b2b3-4541-9895-aa01f774fa7c)


## Changelog

[ui] The Dagster UI allows you to specify extra tags when re-executing
runs from failure from the runs feed re-execute dialog, or by holding
shift when clicking Re-execute menu items throughout the app.

---------

Co-authored-by: bengotow <bgotow@elementl.com>