[dagster-dbt] Update DbtCloudCliInvocation to accept manifest (#28901)
## Summary & Motivation

This PR updated `DbtCloudCliInvocation` to accept manifest objects
instead of the workspace.

This is required for the subsequent PR - when
`DbtCloudWorkspace.cli(...)` is called in the asset decorator, we need
to pass the manifest used to create the asset definitions to the
`DbtCloudJobRunHandler` object, not the manifest of the workspace.
Because of that, we need to pass a manifest object instead of the
workspace object.

Other parameters like `job_id` and `client` also added because we are
not passing the workspace anymore.

## How I Tested These Changes

Updated tests with BK.