[dagster-dbt] Update args with context in DbtCloudWorkspace.cli(..) (#28902)
## Summary & Motivation

This PR uses the asset execution context passed from the asset decorator
to update the cli invocation args in `DbtCloudWorkspace.cli(...)`.

In a subsequent PR, the context will be used downstream to enable
yielding `Outputs` and `AssetCheckResults` when calling the cli
invocation from the asset decorator.

## How I Tested These Changes

Additional tests with BK