Don't modify the unique_id on a CacheableAssetsDefinition when you run it through with_resources
Summary:
Sending this out for discussion. The problem with the current implementation here is that if there is any non-determinism in the resource config, the ID here changes, causing various problems. For example, the dbt cli resource is currently non-deterministic since it sets an absolute path as one of its fields that can change on each deploy.

I can't think of a situation where a CacheableAssetsDefinition would ever need to co-exist with itself with with_resources called on it (unless there is some odd case i'm not thinking of?) so I would expect it to be safe to keep the same unique

TBD based on feedback here

> Insert changelog entry or delete this section.