Include external assets in asset selection filtering (#28951)
## Summary & Motivation

Currently asset selection filtering relies on `useAssetGraphData` in
order to retrieve assets for filtering. The problem is this query doesnt
include external assets. This is intended behavior since we don't want
to show external assets in the asset graph. However, since bringing the
selection syntax over to the catalog table it also means the catalog
can't select external assets. To fix that I'm including external assets
for the asset catalog (which uses `useAssetSelectionFiltering` (along
with alerts and catalog views which should also include these)). The
asset graph uses `useAssetGraphData` directly so it will continue to
exclude external assets

## How I Tested These Changes

locally with app-proxy. Need to come up with jest tests for these but
want to fix this issue for a customer.