Make the unique_id on a CacheableAssetsDefinition when you run it through with_resources more stable (#28939)
## Summary & Motivation
Sending this out for discussion. The problem with the implementation
here is that if there is any non-determinism in the resource config, the
ID here changes on each deploy, causing various problems and race
conditions. For example, the dbt cli resource is currently
non-deterministic since it sets an absolute path as one of its fields
that can change on each deploy.

I can't think of a situation where a CacheableAssetsDefinition would
ever need to co-exist with itself with with_resources called on it
(unless there is some odd case i'm not thinking of?) so I would expect
it to be safe to keep the same unique ID after with_resources is applied
to it.

## How I Tested These Changes
New test case

## Changelog
[dagster-fivetran] Fixed an issue where new runs of code locations using
fivetran assets would sometimes raise a "Failure condition: No metadata
found for CacheableAssetsDefinition" error if the run was started
immediately after a new version of the code location was deployed.