[dagster-fivetran] Move connector selection to state-backed defs (#28955)
## Summary & Motivation

This PR moves the logic to select Fivetran connector using the
connector_selector_fn from `fetch_fivetran_workspace_data()` to
`defs_from_state()` in the state-backed defs.

The main motivation here is that the workspace data should include all
connectors that are active and valid, then we should create definitions
for a subset of connectors based on the logic in `connector_selector_fn`
when users are using the spec loader. This way, the users can create
multiple assets with different `connector_selector_fns` in the same
workspace.

## How I Tested These Changes

Updated tests with BK