[ui] Repair timezone bug in custom date selection (#28967)
## Summary & Motivation

There is a timezone-related issue in the Datepicker-based filter that causes date mismatches when the user's local browser timezone and their Dagster timezone (as set in User Settings) do not match.

For instance, if my local timezone is America/Chicago and my Dagster timezone is Australia/Darwin, and I choose "April 1", react-dates will take this to mean April 1 in Chicago. When I make my selection, however, the Dagster UI will see this timestamp and render it as "April 2", the current date based on my Dagster timezone selection.

To repair this, set the moment timezone globally when opening the datepicker. We currently only use `moment` or `moment-timezone` in this one use case (grepped to check) so I don't want to pull the dependency out into the top-level context provider. But changing it here to match the user's Dagster setting seems to be an appropriate fix.

## How I Tested These Changes

Added Jest specs for timezone behavior.

In the Dagster app, modified my timezone to be Australia/Darwin, then made datepicker selections in the Runs UI. Verified that the tag looks correct, and that the subsequent query has start/end times that match the correct date/time values for these dates in Australia/Darwin.

## Changelog

[ui] Fix custom time datepicker filter selection for users with custom Dagster timezone settings.