Always use graphAssetKeys in `useAssetSelectionFiltering` (#28964)
## Summary & Motivation

I added this logic to side step using the results from selection
filtering if there is no asset selection but overlooked that the
favorites feature in cloud hooks into the selection filtering (via
useAssetGraphData) in order to affect which assets are shown. I will add
a test case in internal for this.

## How I Tested These Changes
used favorites locally.