Remove selection syntax feature flag part 1 (#28969)
## Summary & Motivation

Disable turning off the selection syntax flag for now. In a follow-up PR
i remove all of the code but for now just want to remove the flag so
that new features I'm working on don't need to consider the case when
the
flag is off.

## How I Tested These Changes
BK, open feature flags dialog.