[dg] Default to dataclass in component type generation with an opt-out (#28979)
## Summary & Motivation

While `Component` itself is agnostic to what framework is used to define properties (e.g. they could Pydantic or their own handrolled system) but that shouldn't prevent us from providing an easy path in the default case. So we are going to default to use `@dataclass` but allow an opt-out.

## How I Tested These Changes

BK

## Changelog

NOCHANGELOG