[ui] Marketplace: Build step to pull content from docs (#28978)
## Summary & Motivation

Create a script for copying integrations metadata and markdown information from the `docs` project for reuse in the Dagster app.

- For every `layout: integration` markdown file in the `integrations` section of the `docs` project, extract frontmatter and markdown content into a generated TS file within `ui-core`.
- Additionally, copy over any logo image specified in the frontmatter.
- Update the `MarketplaceHome` and `IntegrationPage` to make use of the generated types.
- Implement text search behavior on `MarketplaceHome`, for now just based on the names of the integrations.

Note that I'm not yet trying to render the markdown into anything special. I'll do that in a followup, since it involves some of the same complexity that `docs` does, with regard to extracting code examples from real python files.

<img width="1290" alt="Screenshot 2025-04-03 at 14 17 50" src="https://github.com/user-attachments/assets/77e9e8fb-d09a-49ce-b0c4-87bfbf837ce4" />

<img width="1011" alt="Screenshot 2025-04-03 at 14 18 57" src="https://github.com/user-attachments/assets/9c6940f8-378d-4d28-9245-e70747900ac6" />

## How I Tested These Changes

`yarn generate-integrationd-docs`, then view Storybook examples.