[ui] Remove remaining uses of RunTable
This PR cleans up a few remaining places where we used the old RunTable component. I manually verified the appearance of the old / new table in each of these callsites, and had to make a few changes to the CSS of `<RunsFeedTable scroll=true />`. That variant was only in use on the backfill runs tab, and I verified that is un-impacted by the CSS changes.

This PR also refines the loading state of the new RunsFeedTable a bit. Since it’s initial launch we added an indeterminate progress bar to the top, and if the result set is empty, autorefresh causes it to transition between empty and loading (with the table headers) every 15s. I made custom empty states render beneath the header so that it does not appear and disappear repeatedly, and adjusted a few empty state borders as necessary.

Screenshots below!