Downgrade logs on CPU quota fetch failures from errors to debug
Summary:
This appears to be an expected case in common scenarios, so make the logger about failing to fetch it much less noisy.

Test Plan:
Confirmed existing test coverage for this case

Changelog:
Fixed an issue where Dagster run metrics produced large amounts of error lines when running in containers without a CPU limit set on the container.