[ui] Remove remaining BackfillTables in favor of RunsFeedTable’s backfills view
This PR moves the remaining BackfillTables (on the job > partition pages) to the Run Feed Table, which can also render a history of backfills. This allows us to remove the old backfill table components, which used slightly different versions of several UI components.

There is one known issue that will result from this change - the new RunsFeedTable cannot filter backfills by code location, so the job backfills page will show results for all jobs with the job name across code locations. To address this, we’ll add a code location filter to the new run feed UI in the future.