Downgrade logs on CPU quota fetch failures from errors to debug (#29008)
## Summary & Motivation
This appears to be an expected case in common scenarios, so make the
logger about failing to fetch it much less noisy.

## How I Tested These Changes
Confirmed existing test coverage for this case

## Changelog
Fixed an issue where Dagster run metrics produced large amounts of error
lines when running in containers without a CPU limit set on the
container.