[dg] Make DgContext work with src-based layout (#29006)
## Summary & Motivation

Prior to this PR, `dg` would crash when working with projects with an
`src`-based layout. This makes it compatible with `src`-layout projects.

The implementation is a little rough-- we just look for packages in both
`src` and in root. The reason for this is that there's not an otherwise
reliable great way to detect `src`-based layout. You could try reading
`pyproject.toml` and looking for config of well-known build tools, but
that's a can of worms. I suspect this solution will work well.

Upstack we are actually going to change the default new project layout
to be src-based.

## How I Tested These Changes

New unit test.

## Changelog

Fixed a bug where `dg` would crash when working with Python packages
with an src-based layout.