[ui] Remove the “legacy runs page” flag and supporting code (#28984)
## Summary & Motivation

This PR removes the feature flag allowing users to fall back to the
legacy runs UI, and removes code that is no longer referenced. There are
still a handful of non-feed `<RunTable />` components, on the
Instigation Tick page, the partition detail panel, the sensor "previous
runs" panel, and the instance concurrency key pool page. At least the
last one is new and can be moved to the new feed table in a follow-up.

Even with those few instances of the old run table remaining, there is a
ton of code we are able to remove now. 🙌 The old run tables were
slightly bespoke in each location, and the new version is a single
component with different filters applied at each callsite.

## How I Tested These Changes

Updated storybooks, ran tests

## Changelog

[ui] The legacy runs page has been removed and is no longer accessible
via a feature flag.

Co-authored-by: bengotow <bgotow@elementl.com>