Make scaffolded schedule include a RunRequest, update tutorial
Summary:
With the current implementation you have to remember to add a RunRequest for it to actually do anything, which is easy to forget, so include it in the scaffolded object.

(Another option would be to make this scaffold a ScheduleDefinition instead of a schedule decorator).

Test Plan: BK, view tutorial docs