Add a "validate_partition_mapping" method on PartitionMapping (#28067)
## Summary ## 
Adds a new method that does basic sanity checking of the upstream and
downstream defs to make sure that they are not incompatible. Can then be
used during definitions loading or "dagster definitions validate" to
verify that all partition mappings are valid. Previously the only way to
get this information was to actually attempt a partition mapping, so it
was cumbersome to check at definition time.

## Test Plan ## 
New test cases