[dg] Refactor PackageEntry types (#28953)
## Summary & Motivation

Internal refactor that gives each package entry an arbitrary list of types. Right now, those types are just scaffold-target and component, but we can add more in the future (e.g. Resolvable)

This is useful for categorizing and filtering entries. Previously, we had a hierarchical system (where ComponentSnap was a subclass of LibraryObjectSnap), which was hard to extend and filter by, as all of these types somewhat overlap with eachother

## How I Tested These Changes

## Changelog

NOCHANGELOG