[dg] Rework `dg list defs` output (#28989)
## Summary & Motivation

This reworks the output of `dg list defs` to be prettier / more in line with `dg list packages`

In service of that goal, I created the bare-minimum "framework" to standardize the pattern both of these use, which is to have a single "outer table" which contains multiple "inner tables". This will of course evolve over time to handle more custom use cases but for now I just wanted to avoid code duplication.

Before:

![Screenshot 2025-04-03 at 5.14.00 PM.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/cTPQ7oSxo9qxXgkzpnaK/437e6f50-03dd-47d6-b58a-c13cc72f839c.png)

After:

![Screenshot 2025-04-03 at 5.09.12 PM.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/cTPQ7oSxo9qxXgkzpnaK/3aad48a4-5fef-4607-81e6-6e901e40c216.png)

## How I Tested These Changes

## Changelog

NOCHANGELOG