[dg] Docs: wrap copy behavior in a try/catch (#29070)
## Summary & Motivation

Add a try/catch to the clipboard write on the dg-docs-components CopyButton, to see if it helps surface issues seen by @schrockn.

Also added an `await` on the `writeText` call to be a bit more correct with subsequent behavior, since the function returns a Promise.

## How I Tested These Changes

`dg docs serve`, verify that copy works as expected on code blocks.