[components] Update components schema to allow declaring env var requirements, validate in check yaml (#28749)
## Summary

Adds a new top-level key to the component file which allows declaring
requirements, including `env` vars (and in the future, resources and
UDFs for the profiles system).

```yaml
type: ...

attributes: ...

requirements:
  env:
    - MY_ENV_VAR
```

This enables us to validate if dependencies are fully satisfied without
fully loading a code location.:
```sh
$ dg check yaml

/Users/ben/repos/components_demo/03-28-env-demo/projects/data_ingest/data_ingest/defs/sling_elt/component.yaml:1 - requirements.env Component uses environment variables that are not specified in the component file: GROUP_NAME
     | 
   1 | type: dagster_sling.SlingReplicationCollectionComponent
     | ^ Component uses environment variables that are not specified in the component file: GROUP_NAME
   2 | attributes:
   3 |   replications:
   4 |   - path: replication.yaml
   5 |     asset_attributes:
   6 |       group_name: "{{ env('GROUP_NAME') }}"
     | 
```

```sh
$ dg check yaml

The following environment variables are used in components but not specified in the .env file or the current shell environment:
GROUP_NAME
```

This behavior can be opted out with the `--no-validate-requirements`
flag.

Similar support could be implemented on `@component`/`@defs_module`: 

```python
@defs_module(requires_env={"MY_ENV_VAR"})
def loader(context: ComponentLoadContext): ...
```

These requirements are validated through `dg check yaml`, which will
warn the user if an env var is not specified in the requirements field
or not provided locally.

#28750 introduces a check yaml flag to automatically fill in these
requirements so this will hopefully not be a big lift from the user's
perspective.

## Test Plan

New unit tests.