[RFC] Definitions.map_asset_specs (#28784)
## Summary & Motivation
Adds a top-level "map_asset_specs" method to Definitions, which takes in an optional asset selection and an AssetSpec -> AssetSpec transformation function.

The idea is to enable last-mile transformation of asset specs on a Definitions object; similar to what `asset_post_processors` does, but for non-components Definitions. In fact, most of this implementation is ripped directly from the asset_post_processors code path, and we reimplement that on top of this functionality.

## How I Tested These Changes
- Additional test_definitions_class.py tests.
## Changelog
- A `Definitions.map_asset_specs` method, which allows for transforming asset specs from any passed in AssetSpec or AssetsDefinition objects which match an asset selection.