Conditionally import dagstermill in toys
We don't explicitly set a dependency on dagstermill in our dagster_test
setup.py. Nor do I really think we should - we do quite a nice job of
keeping the dependencies here quite slim.

But this notebooks job does require it and on a vanilla install, it can
prevent the entire code location from loading.

So let's wrap this in a conditional. I'm open to adding dagstermill in
its own extra or removing this toy job entirely instead.