[dagster-fivetran] Implement FivetranWorkspace.get_or_fetch_workspace_data (#29106)
## Summary & Motivation

This PR follows changes made in response to this
[comment](https://github.com/dagster-io/dagster/pull/28983#discussion_r2031955363).

In hindsight, moving the logic to fetch the workspace data in the
definitions loader seems incorrect. The `FivetranWorkspace` object
should be responsible of handling how to fetch the data, not the
state-backed definitions.

That said, accessing a private method in the definitions loader should
be avoided as initially suggested in the comment.

This PR implements brings back the logic to fetch the workspace data in
`FivetranWorkspace.fetch_fivetran_workspace_data`, and implements
`get_or_fetch_workspace_data` to leverage the state-backed defs in
various internal methods.
- We don't access a private method anymore in the definitions loader
- no breaking change the end users
- state-backed defs are used in the specs loader.

## How I Tested These Changes

Updated tests with BK