[dg] Change sure config from dg.toml -> .dg.toml (#29107)
## Summary & Motivation

Change the user dg config file from `dg.toml` to `.dg.toml` on Unix.
This means that there will be no `dg.toml` file in the home directory
that will be discovered as a containing workspace.

## How I Tested These Changes

Manually, by checking that settings from `~/.dg.toml` were inherited and
those from `~/dg.toml` were not. Used this approach because it's hard to
use an automated test when reading from the home directory.

## Changelog

The `dg` user config file on Unix is now looked for at `~/.dg.toml`
instead of `~/dg.toml`.