Enable Buildkite Test Engine muting (#28971)
This checks the Buildkite API for a list of test cases that have been
"muted" via the Buildkite UI. On every branch but the release branch, it
annotates muted tests with a pytest "skip" annotation. That way, we can
quarantine and triage broken or flaky tests without our build staying
broken while we work on a fix.

Future versions of the buildkite-test-collector plan to support this use
case more directly so we might be able to remove this custom
implementation in the near future.

https://buildkite.com/docs/test-engine/flaky-test-management

A version of this is already running on our internal repo but this sets
it up for this repo before I start evangelizing this functionality more
broadly.