store dbt project dir as a relative path to the cwd instead of an absolute one.
Summary:
in some environments (notably, dagster+ serverless), the working directory can change when the code location is redeployed, since its in a new venv. This means that storing the project dir as an absolute path on the resource runs the risk of an incompatibility after you redeploy. This resolves that by storing the field as a relative path instead.