[ui] Remove legacy RunTable (#28996)
This PR removes the old RunTable, which is now unused. I also traced
through a few components used in RunRow and removed props that were used
to gate the appearance between the old and new versions of the run
table. (useTags is never false, isJob is now always true, etc)

Co-authored-by: bengotow <bgotow@elementl.com>