[asset health] asset check health status calculations (#28795)
## Summary & Motivation
Computes the health status for the asset checks for an asset
Rules:
- `HEALTHY` - the latest completed execution for every check is a
success. Note the latest completed execution for each check may not have
executed based on the most recent materialization of the asset
- `WARNING` - the latest completed execution for any asset check failed
with severity `WARN`(and no checks failed with severity `ERROR`). Note
the latest completed execution for each check may not have executed
based on the most recent materialization of the asset
- `DEGRADED` - the latest completed execution for any asset check failed
with severity `ERROR`. Note the latest completed execution for each
check may not have executed based on the most recent materialization of
the asset
- `UNKNOWN` - any asset checks have never been executed (unless some
checks have failed, then `WARNING` or `DEGRADED` status applies)
- `NOT_APPLICABLE` - the asset has no asset checks defined

tests in https://github.com/dagster-io/internal/pull/14690

## How I Tested These Changes