[asset health] materialization health status calculation (#28816)
## Summary & Motivation
tests in https://github.com/dagster-io/internal/pull/14691

resolver to determine if the latest materialization of an asset passed
or failed.

Follows these rules:
- Partitioned asset
    - `HEALTHY` - all partitions successfully materialized
- `WARNING` - some partitions are successful but any number of
partitions are missing (but no failed partitions)
    - `DEGRADED` - any number of partitions are failed
- `UNKNOWN` - no `partition_stats` for the asset; all partitions are
missing
    - `NOT_APPLICABLE` - no conditions lead to a not applicable status
- Non-partitioned asset
- `HEALTHY` - the latest materialization of an asset was successfully
materialized
    - `WARNING` - no conditions lead to a warning status
    - `DEGRADED` - latest materialization of an asset failed
    - `UNKNOWN` - asset has never had a materialization attempt
    - `NOT_APPLICABLE` - no conditions lead to a not applicable status


- Note: an in progress run to materialize the asset will have no affect
on the status of the asset until a `MATERIALIZATION` or
`ASSET_FAILED_TO_MATERIALIZE` event is emitted from the run

The case when failure events are not stored is based on how we display
status in the UI (see
https://github.com/dagster-io/dagster/blob/3093ad776fdfb8cf7ad52e619f780124ecb5ff30/js_modules/dagster-ui/packages/ui-core/src/asset-graph/AssetNodeStatusContent.tsx#L230
and
https://github.com/dagster-io/dagster/blob/9551a1d6fd0b3571bcac41380d91e9780c917df4/js_modules/dagster-ui/packages/ui-core/src/asset-graph/Utils.tsx#L226-L246)
and running various scenarios in the product and documenting how we
communicate the status of the asset. Because of this, some of the
statuses may not be what you would expect (but are what we currently
show in the UI)! the tests in the linked PR serve mainly as a way to
document the current behavior.

The case when failure events are stored in the UI is a bit easier to
compute, we can just compare the storage ids of the latest successful
and failed materialization events

## How I Tested These Changes