use historic step key vs current step key for asset materialization link (#29087)
## Summary & Motivation
Especially for multi-assets with asset selections, the stepKey can
change across runs and cannot be trusted. This results in broken
experience when using the materialization tag to link to runs.

Should be able to satisfy without doing any extra data fetching
(although it is passing an extra string per node).

## How I Tested These Changes


https://github.com/user-attachments/assets/401a5ce4-5e39-4004-965c-eb18bceb78b2


## Changelog

- [ui] Fixed a bug where the materialization status tag would sometimes
deep link to the wrong step key in the materializing run.