Add retry layer to k8s pipe client consume_pod_logs
Summary:
Retry if there is a transient error while consuming pod logs in the k8s client, using timestamps to prevent the same logs from being logged again.

This does not help with the possibility that the connection can hang. To overcome that we could maybe add a timeout and switch away from streaming to polling, relying on this same scheme to prevent duplicates?

fix pipes tests

> Insert changelog entry or delete this section.