Use containerized s3 in monitoring daemon tests
Most of the complexity here is in getting the docker networking to work
and I chose to do it in a way that was as minimally invasive to how the
test currently works as possible. So the S3 container just rides along
on the existing network. I have medium confidence this will work on
Buildkite.