Modernize postgres docker compose fixtures
This introduces a new set of pytest fixtures on dagster_test for interacting with Postgres docker containers. It allows us to remove a lot of our manual Buildkite networking configuration and should make it easier to extend docker-compose.yml files to add additional services on the correct network.

I'm starting just with the postgres test suite but eventually hope to eliminate all uses of TestPostgresInstance.