Move fixtures into dagster-postgres
dagster-test is a weird package. It has no explicit dependencies but
tons of implicit ones. So it doesn't have dagster-postgres installed
(although they often happen to be installed together in our tox.ini)

Similarly, nothing explicitly depends on it. But almost everything
installs it in its tox.ini.

I think this is because we don't publish the package.

This moves the postgres-specific fixtures into dagster-postgres. But
maintains its implicit dependency on dagster-test.