toys code location for testing asset health (#29190)
## Summary & Motivation
starting a small code location that we can deploy to dogfood to make it
easier to test asset health. Most of our assets in the toys repo always
succeed or always fail, making it hard to test state transitions. We'll
probably need to add some more assets+checks as testing continues, but
this will get us started

## How I Tested These Changes

## Changelog

> Insert changelog entry or delete this section.