Correctly add pytest plugins
I originally had this nested down a level and was bumping into:

https://docs.pytest.org/en/stable/how-to/writing_plugins.html#requiring-plugins-in-non-root-conftests

So I did direct imports.

But sticking it in the conftest.py in the package's root dir (not our
monorepo root dir) satisfies where pytest expects to be able to lookup
these plugins and lets us avoid the nasty ignore/noqa annotations.