[ui] Add Integrations Marketplace to app (#29027)
## Summary & Motivation

Behind a (still-hidden) feature flag, add the Integrations Marketplace to the app.

- Retrieve integration data from our new public API.
- Populate a top-level marketplace list.
- Populate pages for individual integrations.

There is still more to add here that will be in followups:

- More metadata, namely for pypi/repo information and "built by" data.
- The individual integration page needs to be updated to match current mocks.
- We have more work to do on the markdown itself, e.g. showing both `pip` and `uv` commands.

<img width="1570" alt="Screenshot 2025-04-09 at 12 36 50" src="https://github.com/user-attachments/assets/e3f882a8-313c-489c-b109-b7e8bd6242c5" />

<img width="1411" alt="Screenshot 2025-04-09 at 12 39 57" src="https://github.com/user-attachments/assets/da9adc3c-5406-43b1-8955-b61972587385" />

## How I Tested These Changes

Add myself to the feature flag, view Integrations Marketplace. Verify correct rendering of list page and individual integration pages.

Outside of the flag, verify that these routes are not accessible, and that the top nav button link does not appear.