Set a global connect timeout on the fixture
This was working locally but not remotely - I'm just going to stick it
on the fixture itself instead of on the single test that was using it in
hopes that it'll prevent similar issues going forward. If other tests in
the test suite don't like this, I'll move it back to the
test_launch_docker_no_network tests that needed it.