[dbt] fix state_path on DbtCliResource (#29085)
We were not resolving `state_path` as documented when directly setting
on `DbtCliResource`.

I think this went undetected for a while since the more common path is
to set it via `DbtProject` which does the correct resolution.

This was reported as part of #28546, where we are passing to the
resource directly

## How I Tested These Changes

added test

## Changelog

[dagster-dbt] the `state_path` argument to `DbtCliResource` now resolves
relative to the project directory as documented