[dg] Add configuring dg guide (BUILD-820) (#29170)
## Summary & Motivation

Add a guide on "Configuring `dg`". This explains the various kinds of
configuration files and how they interact with CLI options. It also
provides a comprehensive list of settings.

The list of settings was done in an _ad hoc_ way-- in the future I think
we'll want to machine-generate this like API docs, but we need something
out there now.

Note that there was (accidentally) a pre-existing but unlisted page in
`master` called "Configuring `dg`" that was an alternative version of
the current uv integration page. So this PR just edits that and makes it
visible, rather than introducing a new page.

## How I Tested These Changes

Simple tests to make sure the snippets load and validate.