[dg] Add note telling users to fix broken path in components ETL tutorial DbtProjectComponent (#29242)
## Summary & Motivation

The components ETL tutorial currently breaks after generating a DBT
component because the `project_dir` relative path is generated
incorrectly as `../../../dbt/jdbt` when it should be
`../../../../dbt/jdbt`. The necessary path changed after the change to
`src`-based layout.

I'm not sure of how the logic needs to be changed to generate the right
path, but this PR just inserts a note into the tutorial warning users
about this bug so they can fix the path and proceed with the tutorial.
We should remove this when we fix the bug and make the next release.

This was not being caught by our tests because we don't actually load
defs in the tests with the initially generated `component.yaml`, and
subsequent edits silently update the path to the correct
`../../../../dbt/jdbt`.

## How I Tested These Changes

Did the tutorial manually