[1/n]Remove conditional selection syntax flag gating in asset graph. (#29258)
## Summary & Motivation
The feature flag is now always enabled and the ability to turn it off to
be disabled. What's left is removing all of the conditional logic. There
is a lot of it so I'm going to do it over multiple PRs. This is part 1
and removes the selection syntax from the asset graph (and another spot
or two).

## How I Tested These Changes
Just ran local host cloud for this one