[ui] Append a dagster/from_ui tag to every run launched from the UI (#29156)
This is a small change but it applies to:

- Job execution / materialization
- Job backfills
- Asset graph materialization
- Asset graph backfills
- Run re-execution (via extraTags)
- Multiple-run launching
- Evaluating schedules
- Sensor dry runs

## Summary & Motivation

Context:
https://dagsterlabs.slack.com/archives/C058CBEL230/p1743459307738509


https://linear.app/dagster-labs/issue/FE-828/add-a-tag-that-indicates-that-a-run-was-triggered-by-a-user-via-the-ui

Allows filtering for runs created from the UI:

<img width="1079" alt="image"
src="https://github.com/user-attachments/assets/66fb89fa-5107-454a-93ce-d8867a79d1eb"
/>


## How I Tested These Changes

I manually tested most of the workflows above, and I also made changes
to the tests I placed this code very close to the mutations to make sure
it's always sent.

## Changelog

[ui] Runs launched from the Dagster UI get a `dagster/from_ui = true`
tag, making it easy to filter for them.

Co-authored-by: bengotow <bgotow@elementl.com>